requirements.txt
ipython==8.12.3
joblib==1.4.0
numpy==1.26.4
onnxruntime==1.18.0
opencv_contrib_python==4.7.0.72
opencv_python==4.9.0.80
opencv_python_headless==4.8.0.74
pandas==2.2.2
Pillow==10.3.0
plotly==5.22.0
pyrender==0.1.45
pythreejs==2.4.2
PyYAML==6.0.1
PyYAML==6.0.1
reportlab==4.0.9
scikit_learn==1.4.2
streamlit==1.32.2
streamlit_option_menu==0.3.12
trimesh==4.4.0
